requirements.txt
django
django-countries
djangorestframework
django-filter
markdown
pillow
django-jet
gunicorn
pandas
numpy
coreapi
pyyaml
psycopg2-binary
requests
